requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pypdf
sentence-transformers
faiss-cpu
transformers
torch
slowapi